Move employee API into workflow app
